# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book API",
        "description": "API documentation for Laravel Book API",
        "contact": {
            "email": "you@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get list of books",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of books per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for book titles",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by column (title or publication_date)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "title"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc"
                        }
                    },
                    {
                        "name": "genres",
                        "in": "query",
                        "description": "Filter by genre IDs (comma-separated, max 4)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of books",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "genre_id",
                                    "author_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Harry Potter"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Fantasy novel"
                                    },
                                    "genre_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "publication_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-07-08"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{book}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a single book",
                "operationId": "d8bbccd69b954f8a92bd6c3292b15842",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update a book",
                "operationId": "cf6535a3185059974fb8baaf90b29eaf",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "genre_id": {
                                        "type": "integer"
                                    },
                                    "author_id": {
                                        "type": "integer"
                                    },
                                    "publication_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book",
                "operationId": "59b4f85b4a4fe1e69af7004dce3d489f",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        },
...
        "/api/books/{id}/restore": {
            "patch": {
                "tags": [
                    "Books"
                ],
                "summary": "Restore a soft deleted book",
                "operationId": "13cee7f073ef23594e04f8858707fd84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book restored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Book": {
                "title": "Book",
                "required": [
                    "title",
                    "genre_id",
                    "author_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Harry Potter"
                    },
                    "description": {
                        "type": "string",
                        "example": "Fantasy novel"
                    },
                    "genre_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "publication_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2000-07-08"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-21T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-21T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}